Urna_Eletronica/Apuração/requirements_apuracao.txt
python_gnupg==0.5.0
